+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Breaking Changes
 - Environments and organizations have been replaced with namespaces.
+- Removed unused asset metadata field.
 
 ## [2.0.0-beta.7-1] - 2018-10-26